This PR removes the global sort required right before dedup, and redistributes the logic in the proxy heap and async response objects.

We still need to do a global sort in cases where a single store can send multiple replicas of the same series. The querier can detect this case through a warning and it will perform the global sort as before.

* [ ] I added CHANGELOG entry for this change.
* [x] Change is not relevant to the end user.

## Changes

* Move sorting of labels to the AsyncSeriesResponse object, right after a series is received. This allows sorting to be distributed between different goroutines.
* Precalculate a signature for each received series that can be used for fast comparison in the k-way merge algorithm.
* Detect cases where a single store sends multiple replicas of the same series. Emit a warning so that the querier knows it needs to sort the response set.

## Verification

Existing and new unit tests. Also did some manual tests in one of our own environments.

## Benchmark
```
name                                             old time/op  new time/op  delta
QuerySelect/1000000SeriesWith1Samples/select-8    1.13s ± 0%   0.74s ± 0%   ~     (p=1.000 n=1+1)
QuerySelect/100000SeriesWith100Samples/select-8   103ms ± 0%    77ms ± 0%   ~     (p=1.000 n=1+1)
QuerySelect/1SeriesWith10000000Samples/select-8  12.4ms ± 0%  12.8ms ± 0%   ~     (p=1.000 n=1+1)
```

As predicted by the benchmark, I also saw a reduction for a query from 25s to 20s.

Pushing replica labels to the end is still expensive, but as next steps we can do that in stores since replica labels are now added to the `SeriesRequest`
<img width="1881" alt="image" src="https://user-images.githubusercontent.com/1286231/191710255-cdef8064-dca4-4982-a6c0-dfca15b78444.png">