@fpetkovski 2022-09-23
There was one issue with pushdown which should be fixed now. Other CI failures are unrelated to the change.