@fpetkovski 2022-09-29
Here is the alternative implementation which does the sort check only in one place, without relying on responses from stores: https://github.com/thanos-io/thanos/pull/5742. It's simpler to understand and maintain, and results are similar.